Fetcher can download docker image layers
Fetcher supports fetching layers from existing docker images.

This allows a fission deployment to integrate with existing workflows
that build can docker images.

When an archive specifies an "Image" field we attempt to pull that
image from the configured docker registry.

TODO we currently add the docker credentials directly to the environment
of fetcher container. Instead we should insert them as a secret into
the namespace that we're deploying the pod into and reference that
secret using conventional Kubernetes idioms.